added integration of whole system as it stands currently - had to hardocode the random string so teardown could remove the same
