+++ tests/database/test_db_accessor.py
@@ -3,12 +3,32 @@ from redis import Redis
 
 from src.database.db_accessor import REDIS_HOST
 from src.factory import Factory
-from tests.base_test import BaseTest
 
 
-class TestRedisConnection(BaseTest):
+class TestRedisConnection:
+    r: Redis
+
+    @pytest.fixture(scope="function")
+    def _setup_and_teardown_hashed_test_data(self):
+        """
+        - Initialize db connection
+        - Only run tests if reserved keywords are not already in use
+        - Remove all keywords from database after tests run (regardless of pass or fail)
+        """
+        self.r = Redis(REDIS_HOST)
+        self.reserved_hashes = [("urls", "test"), ("users", "user123")]
+        for hash, key in self.reserved_hashes:
+            if self.r.hget(hash, key) is not None:
+                raise AssertionError(f"redis key `{key}` already has data")
+        yield
+        for hash, key in self.reserved_hashes:
+            self.r.hdel(hash, key)
+            self.r.hdel(key, hash)
+
     @pytest.mark.integration_test
     def test_can_add_data(self, _setup_and_teardown_hashed_test_data):
+        print(self.r.hget("urls", "test"))
+        _setup_and_teardown_hashed_test_data
         # Arrange
         db = Factory.create_db_accessor()
 
@@ -28,23 +48,12 @@ class TestRedisConnection(BaseTest):
         assert second_result.status == False
 
     @pytest.mark.integration_test
-    def test_query_with_no_data_returns_false(self, _setup_and_teardown_hashed_test_data):
+    def test_query_with_no_data_returns_false(
+        self, _setup_and_teardown_hashed_test_data
+    ):
         # Arrange
         db = Factory.create_db_accessor()
 
-        result = db.query("url", "test")
+        result = db.query("urls", "test")
 
         assert result.value is None
-
-    @pytest.mark.integration_test
-    def test_can_add_and_query_data(self, _setup_and_teardown_hashed_test_data):
-        # Arrange
-        db = Factory.create_db_accessor()
-
-        # Act
-        db.add("users", "user123", "a user name")
-
-        # Assert
-        result = db.query("users", "user123")
-        assert result.status
-        assert result.value == b"a user name"
